BLD: update pytest ignores to work around pyargs incompatibility
The `--ignore` paths are resolved relative to the current directory
before being compared to the test files if they're not already absolute.
The paths are compared using `fnmatch.fnmatch`, which doesn't treat '/'
specially, so only a single asterisk is needed.

`--rootdir .` is needed so pytest collects the tests from the installed
package instead of `{project}/` (the parent of the configuration file).